requirements.txt
langgraph
langchain-community
streamlit
psycopg2-binary
python-dotenv
ollama